+++ dbt/adapters/flink/constants.py
@@ -1,0 +1,4 @@
+class ExecutionConfig:
+    SAVEPOINT_PATH = "execution.savepoint.path"
+    STATE_PATH = "state.savepoints.dir"
+    JOB_NAME = "pipeline.name"