+++ dbt/adapters/flink/handler.py
@@ -1,10 +1,16 @@
 from datetime import datetime
 from time import sleep
-from typing import Sequence, Tuple, Optional, Any, List
+from typing import Dict, Sequence, Tuple, Optional, Any, List
 
 from dbt.events import AdapterLogger
 
-from dbt.adapters.flink.query_hints_parser import QueryHints, QueryHintsParser, QueryMode
+from dbt.adapters.flink.constants import ExecutionConfig
+from dbt.adapters.flink.query_hints_parser import (
+    QueryHints,
+    QueryHintsParser,
+    QueryMode,
+)
+
 from flink.sqlgateway.client import FlinkSqlGatewayClient
 from flink.sqlgateway.operation import SqlGatewayOperation
 from flink.sqlgateway.result_parser import SqlGatewayResult
@@ -95,21 +101,38 @@ class FlinkCursor:
         logger.debug('Preparing statement "{}"'.format(sql))
         if bindings is not None:
             sql = sql.format(*[self._convert_binding(binding) for binding in bindings])
-        logger.info('Executing statement "{}"'.format(sql))
         self.last_query_hints: QueryHints = QueryHintsParser.parse(sql)
+        execution_config = self.last_query_hints.execution_config
+        start_from_savepoint = False
+        if execution_config:
+            if not self.last_query_hints.test_query:
+                savepoint_path = FlinkJobManager(self.session).stop_job(execution_config)
+                if savepoint_path:
+                    logger.info("f: {}", savepoint_path)
+                    execution_config[ExecutionConfig.SAVEPOINT_PATH] = savepoint_path
+                    start_from_savepoint = True
+            if not start_from_savepoint:
+                logger.info("Job starting without savepoint")
+                execution_config.pop(ExecutionConfig.SAVEPOINT_PATH, None)
+
+        if self.last_query_hints.drop_statement:
+            logger.info("Executing drop statement: {}", self.last_query_hints.drop_statement)
+            FlinkCursor(self.session).execute(self.last_query_hints.drop_statement)
+
         self._set_query_mode()
-        operation_handle = FlinkSqlGatewayClient.execute_statement(self.session, sql)
+        logger.info("Executing statement:\n{}\nExecution config:\n{}", sql, execution_config)
+        operation_handle = FlinkSqlGatewayClient.execute_statement(
+            self.session, sql, execution_config
+        )
         status = self._wait_till_finished(operation_handle)
         logger.info(
-            "Statement executed. Status {}, operation handle: {}".format(
-                status, operation_handle.operation_handle
-            )
+            "Statement executed. Status {}, operation handle: {}",
+            status,
+            operation_handle.operation_handle,
         )
         if status == "ERROR":
             raise Exception("Statement execution failed")
-
         self.last_query_start_time = self._get_current_timestamp()
-
         self.last_operation = operation_handle
 
     def _convert_binding(self, binding):
@@ -184,3 +207,44 @@ class FlinkHandler:
 
     def cursor(self) -> FlinkCursor:
         return FlinkCursor(self.session)
+
+
+class FlinkJobManager:
+    def __init__(self, session: SqlGatewaySession):
+        self.session = session
+
+    def stop_job(
+        self, execution_config: Dict[str, str], with_savepoint: bool = True
+    ) -> Optional[str]:
+        if ExecutionConfig.JOB_NAME not in execution_config:
+            return None
+        job_name = execution_config[ExecutionConfig.JOB_NAME]
+        logger.info("Getting job by name {}", job_name)
+        job_id = self._get_job_id(job_name)
+        if job_id:
+            state_path = execution_config.get(ExecutionConfig.STATE_PATH)
+            logger.info("Stopping job {} using path {}", job_id, state_path)
+            path = self._do_stop_job(job_id, with_savepoint, state_path)
+            logger.info("Job stopped {}", job_id)
+            return path
+        return None
+
+    def _do_stop_job(
+        self, job_id: str, with_savepoint: bool, path: Optional[str] = None
+    ) -> Optional[str]:
+        cursor = FlinkCursor(self.session)
+        hints = f"/** execution_config('{ExecutionConfig.STATE_PATH}={path}') */ " if path else ""
+        savepoint_statement = " WITH SAVEPOINT" if with_savepoint else ""
+        cursor.execute(f"{hints} STOP JOB '{job_id}'{savepoint_statement}")
+        for result in cursor.fetchall():
+            return result[0]
+        return None
+
+    def _get_job_id(self, job_name: str) -> Optional[str]:
+        cursor = FlinkCursor(self.session)
+        cursor.execute("SHOW JOBS")
+        for job in cursor.fetchall():
+            if job_name == job[1]:
+                if job[2] not in ("FAILED", "FINISHED", "CANCELED"):
+                    return job[0]
+        return None