+++ tests/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 
 import os
+
 # import json
 
 # Import the fuctional fixtures as a plugin
@@ -13,10 +14,10 @@ pytest_plugins = ["dbt.tests.fixtures.project"]
 @pytest.fixture(scope="class")
 def dbt_profile_target():
     return {
-        'type': 'flink',
-        'threads': 1,
-        'host': os.getenv('FLINK_SQL_GATEWAY_HOST', '127.0.0.1'),
-        'port': int(os.getenv('FLINK_SQL_GATEWAY_PORT', '8083')),
-        'session_name': os.getenv('SESSION_NAME', 'test_session'),
-        'database': os.getenv('DATABASE_NAME', 'test_db'),
+        "type": "flink",
+        "threads": 1,
+        "host": os.getenv("FLINK_SQL_GATEWAY_HOST", "127.0.0.1"),
+        "port": int(os.getenv("FLINK_SQL_GATEWAY_PORT", "8083")),
+        "session_name": os.getenv("SESSION_NAME", "test_session"),
+        "database": os.getenv("DATABASE_NAME", "test_db"),
     }